Initial design concepts-Possible start to demo files
Initial design attempt with D3:

This design used the HTML and css from the actual housing lottery pages
as a home page. Links were added under the left sidebar menu to possible
designs for revamping the selection of a room.
The 3 links did a course implementation of the following:
1) Bar chart with roll-over tooltips giving building, room type, and
rooms left for selection
2) Zoomable bubble chart of campus areas, building in areas, room types
with available rooms
3) Same as (2) with a tree map.

PaperNotes:
Determining the possible filters and items for selection.
Filter selection and then initial paper mock-up

Demo_Files:
This is the HTML, css and data files that I used for my initial design.
Modify at will.
I am not including StudentSite.css - this file is copyrighted. If there
is missing "stuff" let me know and I will add it to the project2.css.
Otherwise we just use it for the demo